# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: roleId
          in: query
          schema:
            type: ["integer", "null"]
        - name: isActive
          in: query
          schema:
            type: ["boolean", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "409":
          description: Email already exists
...
  /roles/{id}/page-permissions:
    get:
      operationId: getRolePagePermissions
      tags: [roles]
      summary: Get page permissions for a role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Page permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PagePermission"
    put:
      operationId: updateRolePagePermissions
      tags: [roles]
      summary: Update page permissions for a role (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PagePermissionUpdate"
      responses:
        "200":
          description: Updated permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PagePermission"

  # ─── DASHBOARD ──────────────────────────────────────────────────────────────
...
  /dashboard/audit-log:
    get:
      operationId: getAuditLog
      tags: [dashboard]
      summary: Get recent audit log entries
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 50
        - name: userId
          in: query
          schema:
            type: ["integer", "null"]
        - name: action
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"

  # ─── PII ────────────────────────────────────────────────────────────────────
...
  /admin/db-connections:
    get:
      operationId: listDbConnections
      tags: [workflow]
      summary: List all DB connections (admin)
      responses:
        "200":
          description: List of DB connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DbConnection"
    post:
      operationId: createDbConnection
      tags: [workflow]
      summary: Create a DB connection (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DbConnectionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DbConnection"
...
  /workflow/jobs:
    get:
      operationId: listDataJobs
      tags: [workflow]
      summary: List data pipeline jobs
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataJobListResponse"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, firstName, lastName, roleId, roleName, isActive, mfaEnabled, authProvider, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        roleId:
          type: integer
        roleName:
          type: string
        isActive:
          type: boolean
        mfaEnabled:
          type: boolean
        authProvider:
          type: string
          enum: [local, m365]
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
    UserInput:
      type: object
      required: [email, firstName, lastName, roleId, password]
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        roleId:
          type: integer
        password:
          type: string
          minLength: 8
        authProvider:
          type: string
          enum: [local, m365]
          default: local
...
    UserListResponse:
      type: object
      required: [users, total, page, pageSize]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # ─── Roles ───────────────────────────────────────────────────────────────
...
    PagePermission:
      type: object
      required: [id, roleId, pagePath, pageName, canAccess]
      properties:
        id:
          type: integer
        roleId:
          type: integer
        pagePath:
          type: string
        pageName:
          type: string
        canAccess:
          type: boolean
    PagePermissionUpdate:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items:
            type: object
            required: [pagePath, canAccess]
            properties:
              pagePath:
                type: string
              canAccess:
                type: boolean

    # ─── Dashboard ────────────────────────────────────────────────────────────
...
    AuditLogEntry:
      type: object
      required: [id, userId, userEmail, action, details, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        userEmail:
          type: ["string", "null"]
        action:
          type: string
        details:
          type: ["string", "null"]
        ipAddress:
          type: ["string", "null"]
        resourceType:
          type: ["string", "null"]
        resourceId:
          type: ["string", "null"]
        fieldName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AuditLogResponse:
      type: object
      required: [entries, total, page, pageSize]
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AuditLogEntry"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer

    # ─── PII ─────────────────────────────────────────────────────────────────
...
    DbConnection:
      type: object
      required: [id, name, type, host, port, dbName, schemaName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [backoffice, trading]
        host:
          type: string
        port:
          type: integer
        dbName:
          type: string
        schemaName:
          type: string
        fetchQuery:
          type: ["string", "null"]
        outputFilePath:
          type: ["string", "null"]
        createdBy:
          type: ["integer", "null"]
        lastTestedAt:
          type: ["string", "null"]
          format: date-time
        lastTestSuccess:
          type: ["boolean", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DbConnectionInput:
      type: object
      required: [name, type, host, dbName, username, password]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [backoffice, trading]
        host:
          type: string
        port:
          type: integer
          default: 5432
        dbName:
          type: string
        schemaName:
          type: string
          default: public
        username:
          type: string
        password:
          type: string
        outputFilePath:
          type: string
        fetchQuery:
          type: string
...
    DataJob:
      type: object
      required: [id, type, status, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [fetch, upload_csv, push]
        status:
          type: string
          enum: [pending, running, success, failed]
        triggeredBy:
          type: ["integer", "null"]
        triggeredByEmail:
          type: ["string", "null"]
        connectionId:
          type: ["integer", "null"]
        connectionName:
          type: ["string", "null"]
        recordCount:
          type: ["integer", "null"]
        errorMessage:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
          format: date-time
        finishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    DataJobListResponse:
      type: object
      required: [jobs, total, page, pageSize]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/DataJob"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
...